Changed name to OpenMMForceWrapper
