+++ cvpack/__init__.py
@@ -10,12 +10,12 @@ from .attraction_strength import AttractionStrength  # noqa: F401
 from .centroid_function import CentroidFunction  # noqa: F401
 from .composite_rmsd import CompositeRMSD  # noqa: F401
 from .distance import Distance  # noqa: F401
-from .generic import Generic  # noqa: F401
 from .helix_angle_content import HelixAngleContent  # noqa: F401
 from .helix_hbond_content import HelixHBondContent  # noqa: F401
 from .helix_rmsd_content import HelixRMSDContent  # noqa: F401
 from .helix_torsion_content import HelixTorsionContent  # noqa: F401
 from .number_of_contacts import NumberOfContacts  # noqa: F401
+from .openmm_force_wrapper import OpenMMForceWrapper  # noqa: F401
 from .path_in_cv_space import PathInCVSpace  # noqa: F401
 from .radius_of_gyration import RadiusOfGyration  # noqa: F401
 from .radius_of_gyration_sq import RadiusOfGyrationSq  # noqa: F401
@@ -32,12 +32,12 @@ for _cv in [
     CentroidFunction,
     CompositeRMSD,
     Distance,
-    Generic,
     HelixAngleContent,
     HelixHBondContent,
     HelixRMSDContent,
     HelixTorsionContent,
     NumberOfContacts,
+    OpenMMForceWrapper,
     PathInCVSpace,
     RadiusOfGyration,
     RadiusOfGyrationSq,
